rename img/sass/scripts directories
