+++ .stylelintignore
@@ -6,5 +6,5 @@ _/*
 dist/*
 
 # ignore specific files
-src/sass/01-reset/_customreset.scss
-src/sass/01-reset/_normalize.scss
+src/css/01-reset/_customreset.scss
+src/css/01-reset/_normalize.scss
