+++ functions.php
@@ -4,7 +4,7 @@
 define('BLANKET_VERSION', '2.0.1');
 define('BLANKET_THEME', get_stylesheet_directory_uri());
 define('BLANKET_URL', get_bloginfo('url'));
-define('BLANKET_IMG_DIR', BLANKET_THEME . '/src/img/');
+define('BLANKET_IMG_DIR', BLANKET_THEME . '/src/images/');
 define('BLANKET_SVG_DIR', BLANKET_THEME . '/_/svg/');
