+++ package.json
@@ -11,7 +11,7 @@
     "dev": "yarn start",
     "lint": "yarn lint:js; yarn lint:css",
     "lint:eslint": "eslint './src/**/*.js'",
-    "lint:stylelint": "stylelint './src/sass/**/*'",
+    "lint:stylelint": "stylelint './src/css/**/*'",
     "lint:js": "yarn lint:eslint",
     "lint:css": "yarn lint:stylelint",
     "lint:staged": "lint-staged"
