+++ src/css/main.scss
@@ -5,7 +5,7 @@
 @import "04-variables/variables";
 
 body {
-  background-image: url('../../src/img/test.png');
+  background-image: url('../../src/images/test.png');
   background-repeat: no-repeat;
   background-color: orange;
 }
