+++ src/main.js
@@ -1,15 +1,15 @@
 /* blanket front end styles and js */
 
 // import styles for compilation
-import './sass/main';
+import './css/main';
 
 // polyfills
 import 'whatwg-fetch';
 import 'promise-polyfill/src/polyfill';
-import './scripts/polyfills';
+import './javascript/polyfills';
 
 //vendors
 //import '../vendor/noframework.waypoints.js';
 
 //custom js
-import './scripts/app';
+import './javascript/app';
